requirements.txt
BeautifulSoup4
pyyaml
lxml
tqdm
pillow
discord.py
